Add HandlerDecorators registry, so e.g. JDBI-agnostic annotations can have decorator behavior attached to them
